requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
scipy==1.11.4
pillow
streamlit
